backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
pydantic
tenacity
python-dotenv